budget_tracker/requirements.txt
django==3.2
djangorestframework==3.14
django-cors-headers==3.14
djongo==1.3.6
gunicorn==20.1
python-dotenv==1.0
pymongo==3.12.3